+++ Changes.md
@@ -29,7 +29,7 @@ Improvements
   - Reworked output filter handling. Now outputs with different filters can write to the same file (#6574). Also improves tidiness of resulting Arnold scene description.
   - Allow outputs with layer names and outputs without layer names to be merged to the same file.
 - Viewer : Added `Hide`, `Unhide`, and `Prune` menu items to the right-click menu of the 3D view.
-- LightEditor, AttributeEditor :
+- HierarchyView, LightEditor, AttributeEditor :
   - Added `Hide`, `Unhide`, and `Prune` menu items.
   - Added column for displaying and editing scene visibility.
 - LightEditor : Removed `Delete` menu item, use `Prune` instead.